Add guess controller
Add guess.rb to controller and add post 'guess' route.
Currently it works but only when the card is played once.  Need to change the methods returning new cards to be sure they are only working within the constraints of the current round and not globally checking if any guesses ever existed on the card.