Add show round route
Add the route with the stats.  Found a bug in my round logic while trying to look up stats for incorrect guesses.  I think this was in relation to the problem I was seeing earlier where the game was getting stuck in a loop.  My conditional for asking a card if it's available is incorrect.  If you ever give the flash cards an incorrect answer, it never stops returning you a card.  I need to change how it checks for an available card in my models.  Will come back to this tomorroww.