Add progress bar to erb file
Add a new div with a progress bar id.  This gets targeted by the javascript below and assigned a value being passed from the controller.  This the workaround for embedding the ruby into the javascript as opposed to putting the JS in a separate folder like our CSS is.  There is probably a better way to implement this, but I'm not sure how to pass a ruby object to JS besides making a completely new routing call ie: AJAX, which seems like it's unnessicary for just one variable.